notification/main/core/requirements.txt
aio-pika==9.4.1
aiormq==6.8.0
annotated-types==0.7.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
httplib2==0.22.0
idna==3.7
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
pamqp==3.3.0
pika==1.3.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.2
pydantic-settings==2.2.1
pydantic_core==2.18.3
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
